fix: delete tag from core module
